Ditch session for isLoggedIn.
This should mean you can tell the difference between being logged
in or out, or whether we're waiting for the variable to be set.

This still doesn't seem quite right on the login page when you're
logged in and I'm not sure why, but its better than not being able
to login at all.