Mailing list signup
This shows when completing an adventure. Not sure if we want to
add it elsewhere.

If the user is signed in and uses their verified e-mail address
then they just need to click a single button. If they're not
signed in or they chose to use a different e-mail address they
need to complete a captcha and click on a confirmation link
sent by e-mail.